Refine types and provide the correct namespaces
