+++ types/index.d.ts
@@ -1,4 +1,6 @@
 // Reference multiple definitions, so they could act like a bundle
 
-/// <reference path="MiddlewareUtil.d.ts" />
 /// <reference path="ui5_serve.d.ts" />
+/// <reference path="MiddlewareUtil.d.ts" />
+/// <reference path="TaskUtil.d.ts" />
+/// <reference path="MiddlewareParameters.d.ts" />
